requirements.txt
absl-py==2.3.1
amqp==5.3.1
asgiref==3.8.1
astunparse==1.6.3
attrs==25.3.0
billiard==4.2.1
Brotli==1.1.0
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.36.0
colorama==0.4.6
contourpy==1.3.3
crispy-tailwind==1.0.3
cryptography==44.0.2
cycler==0.12.1
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.0.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-debug-toolbar==5.1.0
django-filter==25.1
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-nested-routers==0.94.1
drf-yasg==1.21.10
Faker==33.1.0
fonttools==4.59.0
idna==3.10
inflection==0.5.1
joblib==1.5.1
kiwisolver==1.4.8
kombu==5.5.4
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
namex==0.1.0
numpy==2.1.3
oauthlib==3.2.2
optree==0.17.0
packaging==25.0
pillow==11.1.0
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.2
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.1.0
setuptools==80.9.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
termcolor==3.1.0
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.9.0
wrapt==1.17.2
